Extract the general tab out of the OrganizationDetail component
